+++ docker/terra-node/entrypoint/start.sh
@@ -1,0 +1,5 @@
+#/bin/sh -e
+
+sh /opt/entrypoint/npm_install.sh
+
+npm run start