+++ scripts/aggregate-translations/aggregate-translations.js
@@ -2,21 +2,15 @@ const path = require('path');
 const fse = require('fs-extra');
 const glob = require('glob');
 const supportedLocales = require('./i18nSupportedLocales');
+const Logger = require('../utils/logger');
 
 const aggregateMessages = require('./aggregate-messages');
 const writeAggregatedTranslations = require('./write-aggregated-translations');
 const writeI18nLoaders = require('./write-i18n-loaders');
-
-const defaultSearchPatterns = baseDirectory => ([
-  path.resolve(baseDirectory, 'translations'), // root level translations
-  path.resolve(baseDirectory, 'node_modules', 'terra-*', 'translations'), // root level dependency translations
-  path.resolve(baseDirectory, 'packages', 'terra-*', 'translations'), // package level translations
-  path.resolve(baseDirectory, 'packages', 'terra-*', 'node_modules', 'terra-*', 'translations'), // package level dependency translations
-]);
-
-const customDirectories = (baseDirectory, directories) => (directories.map(dir => path.resolve(baseDirectory, dir)));
+const defaultSearchPatterns = require('./defaultSearchPatterns');
 
 const isFile = filePath => (fse.existsSync(filePath) && !fse.lstatSync(filePath).isDirectory());
+const context = '[terra-toolkit:aggregate-translations]';
 
 const loadConfigFile = (configPath) => {
   if (configPath) {
@@ -41,6 +35,8 @@ const defaults = (options = {}) => {
     fileSystem: options.outputFileSystem || config.outputFileSystem || fse,
     locales: options.locales || config.locales || supportedLocales,
     outputDir: options.outputDir || './aggregated-translations',
+    excludes: options.excludes || config.excludes || [],
+    format: options.format,
   };
 
   if (!defaultConfig.locales.includes('en')) {
@@ -52,18 +48,23 @@ const defaults = (options = {}) => {
 
 const aggregatedTranslations = (options) => {
   const {
-    baseDir, directories, fileSystem, locales, outputDir,
+    baseDir, directories, fileSystem, locales, outputDir, excludes, format,
   } = defaults(options);
 
-  const searchPaths = defaultSearchPatterns(baseDir).concat(customDirectories(baseDir, directories));
+  const searchPaths = [
+    ...defaultSearchPatterns,
+    ...directories,
+  ];
 
   let translationDirectories = [];
   searchPaths.forEach((searchPath) => {
-    translationDirectories = translationDirectories.concat(glob.sync(searchPath));
+    translationDirectories = translationDirectories.concat(glob.sync(searchPath, { cwd: baseDir, ignore: excludes, follow: true }));
   });
 
+  Logger.log(`Aggregating translations for ${Logger.emphasis(locales)} locales.`, { context });
+
   // Aggregate translation messages for each of the translations directories
-  const aggregatedMessages = aggregateMessages(translationDirectories, locales);
+  const aggregatedMessages = aggregateMessages(translationDirectories, locales, fileSystem);
 
   const outputDirectory = path.resolve(baseDir, outputDir);
   fileSystem.mkdirpSync(outputDirectory);
@@ -72,7 +73,7 @@ const aggregatedTranslations = (options) => {
   writeAggregatedTranslations(aggregatedMessages, locales, fileSystem, outputDirectory);
 
   // Write intl and translations loaders for the specified locales
-  writeI18nLoaders(locales, fileSystem, outputDirectory);
+  writeI18nLoaders(locales, fileSystem, outputDirectory, format);
 };
 
 module.exports = aggregatedTranslations;