+++ scripts/aggregate-translations/generate-translation-file.js
@@ -16,24 +16,35 @@ const sortMessages = (messages) => {
   return sortedMessages;
 };
 
-const translationFile = (locale, messages) => (
-  `import { addLocaleData } from 'react-intl';
-import localeData from 'react-intl/locale-data/${locale.split('-')[0]}';
+const translationFile = (locale, baseLocale, messages) => (
+  `'use strict';\n
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.messages = exports.locale = exports.areTranslationsLoaded = undefined;
 
-addLocaleData(localeData);
+var _reactIntl = require('react-intl');
 
-const messages = ${JSON.stringify(messages, null, 2)};
-const areTranslationsLoaded = true;
-const locale = '${locale}';
-export {
-  areTranslationsLoaded,
-  locale,
-  messages
-};`);
+var _${baseLocale} = require('react-intl/locale-data/${baseLocale}');
+
+var _${baseLocale}2 = _interopRequireDefault(_${baseLocale});
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+(0, _reactIntl.addLocaleData)(_${baseLocale}2.default);
+
+var messages = ${JSON.stringify(messages, null, 2)};
+var areTranslationsLoaded = true;
+var locale = '${locale}';
+exports.areTranslationsLoaded = areTranslationsLoaded;
+exports.locale = locale;
+exports.messages = messages;
+`);
 
 const generateTranslationFile = (locale, messages) => {
   const sortedMessages = sortMessages(messages);
-  return translationFile(locale, sortedMessages);
+  const baseLocale = locale.split('-')[0];
+  return translationFile(locale, baseLocale, sortedMessages);
 };
 
 module.exports = generateTranslationFile;