+++ scripts/serve/serve-static-cli.js
@@ -4,15 +4,14 @@ const serve = require('./serve-static');
 const packageJson = require('../../package.json');
 const loadWebpackConfig = require('./loadWebpackConfig');
 
-
 // Parse process arguments
 commander
   .version(packageJson.version)
-  .option('--config <path>', 'The webpack config to serve. Alias for <config>.', undefined)
-  .option('--port <n>', 'The port the app should listen on', parseInt)
-  .option('--host <s>', 'Sets the host that the server will listen on. eg. \'10.10.10.1\'', '0.0.0.0')
+  .option('--config [path]', 'The webpack config to serve. Alias for <config>.', undefined)
+  .option('--port [number]', 'The port the app should listen on', 8080)
+  .option('--host [string]', 'Sets the host that the server will listen on. eg. \'10.10.10.1\'', '0.0.0.0')
   .option('-p, --production', 'Passes the -p flag to the webpack config, if available.')
-  .option('--site <path>', 'The relative path to the static site. This takes precidence over webpack config if both are passed.', undefined)
+  .option('--site [path]', 'The relative path to the static site. This takes precidence over webpack config if both are passed.', undefined)
   .option('--disk', 'The webpack assets will be written to disk instead of a virtual file system.')
   .parse(process.argv);
 
@@ -20,7 +19,7 @@ serve({
   config: loadWebpackConfig(commander.config),
   disk: commander.disk,
   host: commander.host,
-  port: commander.port,
+  port: Number(commander.port),
   production: commander.production,
   site: commander.site,
 });