+++ scripts/serve/serve-static.js
@@ -1,10 +1,14 @@
-/* eslint-disable no-console */
 const express = require('express');
+const fse = require('fs-extra');
 const path = require('path');
 const webpack = require('webpack');
 const clone = require('clone');
 const MemoryFS = require('memory-fs');
 const mime = require('mime-types');
+
+const Logger = require('../utils/logger');
+
+const context = '[Terra-Toolkit:serve-static]';
 
 // Run webpack on the provided webpack config and save to either the virtual file system or disk.
 const compile = (webpackConfig, disk) => (
@@ -14,16 +18,16 @@ const compile = (webpackConfig, disk) => (
     if (!disk) {
       compiler.outputFileSystem = new MemoryFS();
     }
-    console.log('[Terra-Toolkit:serve-static] Webpack compilation started');
+    Logger.log('Starting Webpack compilation', { context });
     compiler.run((err, stats) => {
       if (err || stats.hasErrors()) {
-        console.log('[Terra-Toolkit:serve-static] Webpack compiled unsuccessfully');
-        reject(err || new Error(stats.toJson().errors));
+        Logger.log(`Webpack failed to compile in ${webpackConfig.mode} mode`, { context });
+        reject(err || Logger.error((stats.toJson().errors)));
       } else {
         if (stats.hasWarnings()) {
-          console.warn(stats.toJson().warnings);
+          Logger.warn(stats.toJson().warnings, { context });
         }
-        console.log('[Terra-Toolkit:serve-static] Webpack compiled successfully');
+        Logger.log(`Webpack compiled successfully in ${Logger.emphasis(webpackConfig.mode.toUpperCase())} mode`, { context });
         resolve([webpackConfig.output.path, compiler.outputFileSystem]);
       }
     });
@@ -34,7 +38,13 @@ const compile = (webpackConfig, disk) => (
 const generateSite = (site, config, disk, production) => {
   if (site) {
     const sitePath = path.join(process.cwd(), site);
-    return Promise.resolve([sitePath, undefined]);
+
+    if (fse.existsSync(sitePath) && fse.lstatSync(sitePath).isDirectory()) {
+      const fileSystem = disk ? fse : undefined;
+      return Promise.resolve([sitePath, fileSystem]);
+    }
+
+    return Promise.reject(Logger.error(`Could not serve static site from ${sitePath}.`, { context }));
   }
 
   if (config) {
@@ -52,7 +62,7 @@ const generateSite = (site, config, disk, production) => {
     return compile(webpackConfig, disk);
   }
 
-  return Promise.reject(new Error('[Terra-Toolkit:serve-static] No webpack configuration provided.'));
+  return Promise.reject(Logger.error('No webpack configuration provided.', { context }));
 };
 
 // Set the test locale for the html file
@@ -70,7 +80,6 @@ const setSiteLocale = (fileContent, locale) => {
 
   return content.replace(/<html/, `<html ${langLocale}`);
 };
-
 
 // Setup an app server that reads from a filesystem.
 const virtualApp = (site, index, locale, fs, verbose) => {
@@ -110,7 +119,7 @@ const virtualApp = (site, index, locale, fs, verbose) => {
 
   if (verbose) {
     app.use((req, res, next) => {
-      const err = new Error(`Not Found: ${req.originalUrl}`);
+      const err = Logger.error(`Not Found: ${req.originalUrl}`, { context });
       err.status = 404;
       next(err);
     });
@@ -160,7 +169,7 @@ const serve = (options) => {
     .then(([sitePath, fs]) => serveSite(sitePath, fs, appIndex, appLocale, verbose))
     .then((app) => {
       const server = app.listen(appPort, host);
-      console.log(`[Terra-Toolkit:serve-static] Server started listening at port:${appPort}`);
+      Logger.log(`Server started listening at port:${appPort}`, { context });
       return server;
     });
 };