+++ src/wdio/services/TerraCommands/visual-regression.js
@@ -1,3 +1,5 @@
+import Logger from '../../../../scripts/utils/logger';
+
 /**
 * Helper method to determine the screenshot tag name, the element selector, the viewport(s)
 * in which to take the screenshots, as well as the capture screenshot options to be passed
@@ -43,6 +45,10 @@ const determineScreenshotOptions = (...args) => {
 * @property {Array} args - An object containing the CSS custom properties to assert.
 */
 const themeEachCustomProperty = (...args) => {
+  if (global.browser.options.terra.disableThemeTests) {
+    return;
+  }
+
   // If more than 1 argument, selector is first
   const selector = args.length > 1 ? args[0] : global.browser.options.terra.selector;
   // Style properties are always last.
@@ -62,11 +68,15 @@ const themeEachCustomProperty = (...args) => {
 * @property {Array} args - An object containing the options for themeCombinationOfCustomProperties and  CSS custom properties to assert.
 */
 const themeCombinationOfCustomProperties = (...args) => {
+  if (global.browser.options.terra.disableThemeTests) {
+    return;
+  }
+
   const selector = args[0].selector ? args[0].selector : global.browser.options.terra.selector;
   const styleProperties = args[0].properties ? args[0].properties : [];
 
   if (!args[0].testName) {
-    throw new Error(`A test name for themeCombinationOfCustomProperties test is not provided.
+    throw Logger.error(`A test name for themeCombinationOfCustomProperties test is not provided.
 A testName property should be set in the options object passed to the themeCombinationOfCustomProperties to uniquely identify it.`);
   }