+++ src/wdio/services/TerraService.js
@@ -2,6 +2,7 @@ import chai from 'chai';
 import chaiMethods from './TerraCommands/chai-methods';
 import accessiblity from './TerraCommands/accessiblity';
 import visualRegression from './TerraCommands/visual-regression';
+import Logger from '../../../scripts/utils/logger';
 import SERVICE_DEFAULTS from '../../../config/wdio/services.default-config';
 
 const { terraViewports: VIEWPORTS } = SERVICE_DEFAULTS;
@@ -29,7 +30,7 @@ const setViewport = (formFactor) => {
     if (terraViewport !== undefined && typeof terraViewport === 'object') {
       global.browser.setViewportSize(terraViewport);
     } else {
-      throw new Error('The formFactor supplied is not a Terra-defined viewport size.');
+      throw Logger.error('The formFactor supplied is not a Terra-defined viewport size.', { context: '[Terra-Toolkit:terra-service]' });
     }
   }
 };
@@ -43,7 +44,6 @@ const setViewport = (formFactor) => {
 export default class TerraService {
   // eslint-disable-next-line class-methods-use-this
   before() {
-    setViewport(global.browser.options.formFactor);
     chai.config.showDiff = false;
     global.expect = chai.expect;
     global.should = chai.should();
@@ -58,5 +58,10 @@ export default class TerraService {
     };
     chai.Assertion.addMethod('accessible', chaiMethods.accessible);
     chai.Assertion.addMethod('matchReference', chaiMethods.matchReference);
+    // IE driver takes a longer to be ready for browser interactions
+    if (global.browser.desiredCapabilities.browserName === 'internet explorer') {
+      global.browser.pause(10000);
+    }
+    setViewport(global.browser.options.formFactor);
   }
 }