+++ tests/jest/write-i18n-loaders.test.js
@@ -6,12 +6,12 @@ const writeI18nLoaders = require('../../scripts/aggregate-translations/write-i18
 
 global.console = { warn: jest.fn() };
 
-const locales = ['en', 'es'];
+const locales = ['en', 'en-US', 'es'];
 const memoryFS = new MemoryFileSystem();
 const testFileSystems = { fse, memoryFS };
 
 Object.keys(testFileSystems).forEach((testFS) => {
-  describe(`write i18n loaders for ${testFS} fileSystem`, () => {
+  describe(`write compiled i18n loaders for ${testFS} fileSystem`, () => {
     const fileSystem = testFileSystems[testFS];
     const outputDir = '/aggregated-translations';
     let writtenFilePaths;
@@ -27,7 +27,7 @@ Object.keys(testFileSystems).forEach((testFS) => {
       console.warn.mockClear();
     });
 
-    it('writes loader files', () => {
+    it('writes compiled loader files', () => {
       const outputFiles = [
         path.resolve(process.cwd(), outputDir, 'intlLoaders.js'),
         path.resolve(process.cwd(), outputDir, 'translationsLoaders.js'),
@@ -38,6 +38,21 @@ Object.keys(testFileSystems).forEach((testFS) => {
       expect(writtenFilePaths).toEqual(outputFiles);
       expect(writtenFiles[0]).toMatchSnapshot();
       expect(writtenFiles[1]).toMatchSnapshot();
+      expect(writtenFiles[2]).toMatchSnapshot();
+    });
+
+    it('writes compiled loader files with modern format if specified', () => {
+      const outputFiles = [
+        path.resolve(process.cwd(), outputDir, 'intlLoaders.js'),
+        path.resolve(process.cwd(), outputDir, 'translationsLoaders.js'),
+      ];
+
+      writeI18nLoaders(locales, fileSystem, outputDir, 'es6');
+
+      expect(writtenFilePaths).toEqual(outputFiles);
+      expect(writtenFiles[0]).toMatchSnapshot();
+      expect(writtenFiles[1]).toMatchSnapshot();
+      expect(writtenFiles[2]).toMatchSnapshot();
     });
 
     it('logs a warning message if a locale is not a terra-supported locale', () => {
@@ -54,6 +69,7 @@ Object.keys(testFileSystems).forEach((testFS) => {
 
       expect(writtenFiles[0]).toMatchSnapshot();
       expect(writtenFiles[1]).toMatchSnapshot();
+      expect(writtenFiles[2]).toMatchSnapshot();
     });
   });
 });